Devops/ci cd (#1)
* Adds CI workflow with linting, testing, and build

Sets up a comprehensive CI pipeline using GitHub Actions.

This workflow includes linting to maintain code quality,
testing to ensure code reliability, and building to prepare
the application for deployment.

* Updates dependencies in yarn.lock

Updates several dependencies, including eslint-related packages and their sub-dependencies, in the yarn.lock file to align with newer versions and address potential security vulnerabilities.

* Removes yarn.lock from .gitignore

Removes yarn.lock from the .gitignore file.

This allows the yarn.lock file to be tracked in the repository.
This ensures that all developers are using the same versions of dependencies, leading to more consistent builds and fewer dependency-related issues.

* Updates ESLint and related dependencies

Updates ESLint and its plugins to the latest versions to improve code linting and maintainability.

This update includes several new ESLint plugins and configurations, enhancing code quality and consistency. It also resolves compatibility issues between ESLint and other dependencies.

* Adds GitHub Pages deployment workflow

Sets up a workflow to automatically deploy the application to GitHub Pages upon successful completion of the CI workflow on the main branch.

This automates the deployment process, ensuring that the live GitHub Pages site is always up-to-date with the latest changes from the main branch.

The workflow includes steps for checking out the repository, configuring GitHub Pages, downloading the build artifact, uploading the artifact to GitHub Pages, and deploying to GitHub Pages.

* Adds deploy workflow and artifact naming

Adds a workflow to deploy the application to GitHub Pages upon successful CI runs on specified branches.

Improves artifact naming by dynamically generating names based on the commit message for better traceability. This helps in identifying the specific build associated with a commit.

* Sanitizes commit message for artifact name

Ensures commit messages used as artifact names are safe by:

- Replacing spaces with dashes.
- Limiting to alphanumeric, dots, underscores, and dashes.
- Truncating to a maximum length of 100 characters.
- Providing a fallback name if the sanitized message is empty.