Feature/upload page (#3)
* Restructures project for better organization

Refactors the project's directory structure to improve
maintainability and scalability.

Moves and renames files to align with a feature-based
organization, grouping related components, hooks, and
services together. Also creates new directories for assets
and styles to centralize resources.

* WIP:- Migrates to webpack for improved build process

Switches from react-scripts to webpack for better control and optimization of the build process.

Configures babel, eslint, and jest for seamless integration with webpack.
Updates CI workflow to align with webpack build process.

This migration allows for greater flexibility in configuring the build pipeline and optimizing for production.

* WIP:- Migrates to webpack for improved build process

Switches from react-scripts to webpack for better control and optimization of the build process.

Configures babel, eslint, and jest for seamless integration with webpack.
Updates CI workflow to align with webpack build process.

This migration allows for greater flexibility in configuring the build pipeline and optimizing for production.